% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{opinosis}
\alias{opinosis}
\title{opinosis dataset of 51 user reviews with topics and five summaries. each topic contains several sentences from different user reviews}
\format{A data.frame: \describe{a data.frame of 51 topics and summaries}}
\usage{
data("opinosis")
}
\description{
opinosis dataset of 51 user reviews with topics and five summaries. each topic contains several sentences from different user reviews
}
\references{
\url{http://kavita-ganesan.com/opinosis-opinion-dataset/#.XOyPdYhKhPZ}
}
\keyword{datasets}
