% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stopwords_longlist}
\alias{stopwords_longlist}
\title{a very long list of english stopwords}
\format{A character vector: \describe{a vector of stopwords}}
\usage{
data("stopwords_longlist")
}
\description{
caution: check the list to make sure any domain-specific words are not being used as stopwords
}
\keyword{datasets}
